Separate user input from the foundational wrapper
