+++ client/engine/backend/foundational_wrapper.py
@@ -1,4 +1,4 @@
-from typing import Any, List, Tuple
+from typing import Any
 
 import pygame
 
@@ -24,15 +24,6 @@ class FoundationalClock:
 
 
 class FoundationalWrapper:
-    K_RETURN = pygame.K_RETURN
-    K_ESCAPE = pygame.K_ESCAPE
-    K_BACKSPACE = pygame.K_BACKSPACE
-    KEYDOWN = pygame.KEYDOWN
-    MOUSEBUTTONDOWN = pygame.MOUSEBUTTONDOWN
-
-    @staticmethod
-    def get_event() -> List[pygame.event.Event]:
-        return pygame.event.get()
 
     @staticmethod
     def quit() -> None:
@@ -51,10 +42,6 @@ class FoundationalWrapper:
     @staticmethod
     def get_clock_ticks() -> int:
         return int(pygame.time.get_ticks())
-
-    @staticmethod
-    def get_mouse_position() -> Tuple[int, int]:
-        return pygame.mouse.get_pos()
 
     @staticmethod
     def set_mouse_cursor(new_cursor: str) -> None:
