+++ client/engine/features/user_input/keyboard.py
@@ -1,14 +1,14 @@
 from typing import Any, List
 
-from client.engine.backend.foundational_wrapper import FoundationalWrapper
+from client.engine.backend.input import InputBackend
 
 
 class KeyboardInput:
     def _get_key_value(self, event: Any) -> Any:
         special_keys = {
-            FoundationalWrapper.K_RETURN: "return",
-            FoundationalWrapper.K_ESCAPE: "escape",
-            FoundationalWrapper.K_BACKSPACE: "backspace",
+            InputBackend.K_RETURN: "return",
+            InputBackend.K_ESCAPE: "escape",
+            InputBackend.K_BACKSPACE: "backspace",
         }
         if event.key in special_keys:
             return special_keys[event.key]
@@ -16,7 +16,5 @@ class KeyboardInput:
             return event.unicode
 
     def read(self, events: List[Any]) -> List[str]:
-        keydowns = [
-            event for event in events if event.type == FoundationalWrapper.KEYDOWN
-        ]
+        keydowns = [event for event in events if event.type == InputBackend.KEYDOWN]
         return [self._get_key_value(event) for event in keydowns]
